work on skeletal rendering and fixed bf3d structs

+++ src/Graphics/Hierarchy.hpp
@@ -37,8 +37,8 @@ namespace anvil
 		inline int GetPivotCount() { return m_pivotCount; }
 		inline void SetCenterPos(glm::vec3 centerPos) { m_centerPos = centerPos; }
 		inline glm::vec3 GetCenterPos() { return m_centerPos; }
-		inline void AddParentID(std::uint16_t ID) { m_parentIDs.push_back(ID); }
-		inline std::vector<std::int16_t> GetParentIDs() { return m_parentIDs; }
+		inline void AddParentID(std::int32_t ID) { m_parentIDs.push_back(ID); }
+		inline std::vector<std::int32_t> GetParentIDs() { return m_parentIDs; }
 		inline void AddPivot(glm::f32mat4x4 pivot) { m_pivots.push_back(pivot); };
 		inline std::vector<glm::f32mat4x4> GetPivots() { return m_pivots; }
 
@@ -46,7 +46,7 @@ namespace anvil
 		std::string m_name;
 		int m_pivotCount;
 		glm::vec3 m_centerPos;
-		std::vector<std::int16_t> m_parentIDs;
+		std::vector<std::int32_t> m_parentIDs;
 		std::vector<glm::f32mat4x4> m_pivots;
 	};
 }
